Added edit mode layout with markdown editor.
If an existing article is specified then it will be display with a
markdown rendered view. If edit mode is set then an input for title and
textarea for content is displayed. If an article that doesn't exist is
requested then a new article is created and edit mode is entered
automatically.